requirements.txt
streamlit
python-dotenv
google-generativeai
networkx
plotly
